Review `Processor`

+++ src/applications/pixel_war/mod.rs
@@ -10,4 +10,4 @@ pub use paint::Paint;
 pub use processor::Processor;
 pub use processor_settings::ProcessorSettings;
 
-pub const CANVAS_EDGE: u16 = 2048;
+pub const CANVAS_EDGE: usize = 2048;
